+++ src/tests.rs
@@ -1,25 +1,26 @@
 use crate::{Backend, TestCommand};
 use prettycli::info;
-use std::ffi::OsString;
 use std::os::unix::process::CommandExt;
 use std::process::Command;
 
 pub struct Tests {
     backend: Option<Backend>,
-    passthrough: Option<Vec<OsString>>,
+    passthrough: Vec<String>,
 }
 
 impl Tests {
+    /// Creates a Test struct
     pub fn new(test: TestCommand) -> Tests {
         let backend = test.backend.map(Backend::load);
 
         Tests {
             backend,
-            passthrough: test.passthrough,
+            passthrough: test.args,
         }
     }
 
-    pub(crate) fn test(&self) {
+    /// Runs a test suite
+    pub fn test(&self) {
         match &self.backend {
             None => {
                 self.test_mock_vm();
@@ -34,7 +35,10 @@ impl Tests {
         self.backend.clone().unwrap().build();
 
         let mut test_args = self.get_test_args();
-        test_args.append(&mut vec!["--no-default-features", "--features=wasm-test"]);
+        test_args.append(&mut vec![
+            "--no-default-features".to_string(),
+            "--features=wasm-test".to_string(),
+        ]);
 
         info("Running cargo test...");
         Command::new("cargo").args(test_args).exec();
@@ -42,20 +46,21 @@ impl Tests {
 
     fn test_mock_vm(&self) {
         let mut test_args = self.get_test_args();
-        test_args.append(&mut vec!["--no-default-features", "--features=mock-vm"]);
+        test_args.append(&mut vec![
+            "--no-default-features".to_string(),
+            "--features=mock-vm".to_string(),
+        ]);
+
+        info("Running cargo test...");
         Command::new("cargo").args(test_args).exec();
     }
 
-    fn get_test_args(&self) -> Vec<&str> {
-        let mut test_args = vec!["test"];
-        match &self.passthrough {
-            None => {}
-            Some(passthrough) => {
-                let passthrough = passthrough.first().unwrap().as_os_str().to_str().unwrap();
-                let mut vec: Vec<&str> = passthrough.split(' ').collect();
-                test_args.append(&mut vec);
-            }
+    fn get_test_args(&self) -> Vec<String> {
+        let mut test_args = vec!["test".to_string()];
+        if !self.passthrough.is_empty() {
+            test_args.append(&mut self.passthrough.clone());
         }
+
         test_args
     }
 }